WIP: new storage using bolt backend
